pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "azure-provisioner"
version = "0.1.0"
description = "Quota-aware Bicep generator and region selector for Azure deployments"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "John Lam", email = "jflam@microsoft.com"},
]
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.2",
    "azure-identity>=1.13.0",
    "azure-mgmt-quota>=1.0.0",
    "rich>=13.4.2",
    "typer>=0.9.0",
    "tenacity>=8.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
provisioner = "main:app"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"

[tool.hatch.build.targets.wheel]
packages = ["provisioner"]
